% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{apply_model}
\alias{apply_model}
\title{a quick function based on the original code}
\usage{
apply_model(df, model, threads = 1, clean_names = TRUE)
}
\arguments{
\item{df:}{A data.frame}
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
