% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{apply_pipeline}
\alias{apply_pipeline}
\title{Applies a list of functions to the given input in order}
\usage{
apply_pipeline(data, pipeline_functions)
}
\arguments{
\item{data:}{the input data.}

\item{functions:}{a vector or list of functions.
Each must take exactly one argument.}
}
\value{
explanation
}
\description{
Long Description here
}
\examples{
Not Yet Implmented

}
\seealso{
None
}
