% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{calc_num_tiles}
\alias{calc_num_tiles}
\title{removes empty rows from the data.frame}
\usage{
calc_num_tiles(file_path, max_size = 128)
}
\arguments{
\item{file_path:}{a string}
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
