% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{calculate_veg_index}
\alias{calculate_veg_index}
\title{Calculates the vegetation indices for the input dataframe}
\usage{
calculate_veg_index(x)
}
\value{
explanation
}
\description{
Long Description here
}
\examples{
Not Yet Implmented

}
\seealso{
None
}
