% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{clean_df_colnames}
\alias{clean_df_colnames}
\title{Cleans the column names of all spaces
and other punctuation}
\usage{
clean_df_colnames(df)
}
\arguments{
\item{df:}{data.frame to have column names modified}
}
\value{
a dataframe with the same contents 
but modified column neams
}
\description{
Long Description here
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
