% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{clean_names}
\alias{clean_names}
\title{a quick function based on the original code}
\usage{
clean_names(variables)
}
\arguments{
\item{df:}{A data.frame}
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
