% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{convert_and_save_output}
\alias{convert_and_save_output}
\title{aconverts the output to the correct data type (base::data.frame or raster::rasterLayer) and save it to disk}
\usage{
convert_and_save_output(df, save_path = NULL, return_raster = TRUE)
}
\arguments{
\item{df:}{A data.frame}

\item{save_path:}{the file path for saving the converted file.  If NULL (default), no file is saved.}

\item{return_raster:}{If TRUE, the function converts 'df' to a raster before saving and returning the converted data. 
If FALSE, the data.frame is saved to disk, but no coversion is made.}
}
\value{

}
\description{
In addition to converting the file, it saves the file to disk.  
In order to avoid issues with overwriting files, the function will 
adjoin 16 random characters to the filename to avoid collision.
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
