% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{convert_fg1_int}
\alias{convert_fg1_int}
\title{a quick function based on the original code}
\usage{
convert_fg1_int(df)
}
\arguments{
\item{df:}{A data.frame}
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
