% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{convert_fg2_int}
\alias{convert_fg2_int}
\title{aconverts functional group 2 codes from integers to Strings}
\usage{
convert_fg2_int(df)
}
\arguments{
\item{df:}{A data.frame.  Assumes that the target data is in column 'z'}
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
