% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{convert_species_int}
\alias{convert_species_int}
\title{converts the species integer codes to the species name (strings)}
\usage{
convert_species_int(df)
}
\arguments{
\item{df:}{A data.frame}
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
