% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{convert_species_string}
\alias{convert_species_string}
\title{converts the species names on the}
\usage{
convert_species_string(df)
}
\arguments{
\item{df:}{A data.frame}
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
