% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{datacube_to_csv}
\alias{datacube_to_csv}
\title{a quick function based on the original code}
\usage{
datacube_to_csv(raster_path, save_path)
}
\arguments{
\item{df:}{A data.frame}
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
