% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{deriv_combine}
\alias{deriv_combine}
\title{COmbines derivatives (veg indices) and predictions}
\usage{
deriv_combine(x, veg_index)
}
\arguments{
\item{filepath}{(optional): A parallel computing framework}
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented

}
\seealso{
None
}
