% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{df_to_speclib}
\alias{df_to_speclib}
\title{Converts Data Frame to a spaectral library}
\usage{
df_to_speclib(df, type = "hsdar")
}
\arguments{
\item{df:}{A dataframe to convert}
}
\value{
A spectral library
}
\description{
Converts Data Frame to a spaectral library
}
