% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{.calculate_veg_index_from_df}
\alias{.calculate_veg_index_from_df}
\title{Calculates the Vegetation Index (via hsdar package)}
\usage{
.calculate_veg_index_from_df(
  df,
  indices = hsdar::vegindex(),
  aviris = c(-58),
  headwall = NULL
)
}
\arguments{
\item{df:}{A dataframe of spectral data}

\item{indices:}{(optional) base vegetation index for the calculations.
If none is supplied, one will be calculated using the hsdar package}

\item{aviris:}{The spectral indices of interest from the AVIRIS data
(default list of only -58)}

\item{headwall:}{The headwall indeices (bands) of interest.
Defaults to NULL, in which case the default bands will be used:
-{3,26,27,31,32,33,35,48,49,58,60,66,67,71,82,99,102,103,104,105}}
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented
}
\seealso{
hsdar::vegindex()
}
