% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{.calculate_veg_index_from_speclib}
\alias{.calculate_veg_index_from_speclib}
\title{One Line}
\usage{
.calculate_veg_index_from_speclib(speclib)
}
\value{
explanation
}
\description{
Long Description here
}
\examples{
Not Yet Implmented

}
\seealso{
None
}
