% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{.convert_tile_filename}
\alias{.convert_tile_filename}
\title{a quick function based on the original code}
\usage{
.convert_tile_filename(tile_path, config)
}
\arguments{
\item{df:}{A data.frame}
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
