% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{estimate_land_cover}
\alias{estimate_land_cover}
\title{equivalent to the old LandCoverEstimator()}
\usage{
estimate_land_cover(
  input_filepath,
  config_path = "./config.json",
  cache_filepath = "./",
  output_filepath = "./predictions.grd",
  use_external_bands = FALSE
)
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented

}
\seealso{
None
}
