% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{extract_bands}
\alias{extract_bands}
\title{Extracts indices from data frame column names}
\usage{
extract_bands(df)
}
\arguments{
\item{df:}{A dataframe containing spectral data}
}
\value{
explanation
}
\description{
Long Description here
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
