% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{filter_bands}
\alias{filter_bands}
\title{a quick function based on the original code}
\usage{
filter_bands(df)
}
\arguments{
\item{df:}{A data.frame}
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
