% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{filter_empty_points}
\alias{filter_empty_points}
\title{removes empty rows from the data.frame}
\usage{
filter_empty_points(df)
}
\arguments{
\item{df:}{A data.frame}
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
