% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{get_vegetation_indices}
\alias{get_vegetation_indices}
\title{Calculates AVIRIS Vegetation Indices}
\usage{
get_vegetation_indices(df, ml_model, cluster = NULL)
}
\arguments{
\item{cluster}{(optional): A parallel computing framework}

\item{base_index}{(optional): vegetation indices for computations.}
}
\value{
vegetation indices for supplied
}
\description{
Long Description here
}
\examples{
Not Yet Implmented

}
\seealso{
None
}
