% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{impute_spectra}
\alias{impute_spectra}
\title{Lone line explanation}
\usage{
impute_spectra(x, cluster = NULL, method = "missForest")
}
\arguments{
\item{x:}{a dataframe to be imputed}
}
\value{
a dataframe with missing values filled
}
\description{
Long Description here
}
\examples{
Not Yet Implmented

}
\seealso{
None
}
