% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{load_config}
\alias{load_config}
\title{Function Reads in the configuration file}
\usage{
load_config(filename)
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented

}
\seealso{
None
}
