% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{load_csv}
\alias{load_csv}
\title{Lone line explanation}
\usage{
load_csv(filepath, output_type = "df")
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented

}
\seealso{
None
}
