% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{load_model}
\alias{load_model}
\title{Function Reads in the data and replace/removes weird values}
\usage{
load_model(filepath)
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented

}
\seealso{
None
}
