% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{make_speclib_derivs}
\alias{make_speclib_derivs}
\title{Function Reads in the data and replace/removes weird values}
\usage{
make_speclib_derivs(filename)
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented

}
\seealso{
None
}
