% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{make_tiles}
\alias{make_tiles}
\title{creates tiles from raster brick and returns the locations on disk}
\usage{
make_tiles(
  raster_obj,
  num_x = 10,
  num_y = 10,
  save_path = "./",
  cluster = NULL,
  verbose = FALSE
)
}
\arguments{
\item{raster_obj:}{a rasterBrick object (RasterLayer and RasterStack should also work but are not tested)}

\item{num_x:}{number of tiles in x direction}

\item{num_y:}{number of tiles in y direction}

\item{save_path:}{folder to save the tiles}

\item{cluster:}{parallel compute cluster (e.g. from parallel::beginCluster()), or NULL}

\item{verbose:}{(boolean, default: false) determines whether details are printed during evaluation}
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented

}
\seealso{
None
}
