% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{merge_tiles_gdal}
\alias{merge_tiles_gdal}
\title{removes empty rows from the data.frame}
\usage{
merge_tiles_gdal(
  target_files,
  output_filename,
  cache_vrt = "./temp_cache.vrt",
  return_raster
)
}
\arguments{
\item{df:}{A data.frame}
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
