% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{plot_agg_results}
\alias{plot_agg_results}
\title{Lone line explanation}
\usage{
plot_agg_results(df, save_file = "Output/results.jpeg")
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented

}
\seealso{
None
}
