% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{postprocess_predictions}
\alias{postprocess_predictions}
\title{Performs post-processing for the process_tile function}
\usage{
postprocess_predictions(predictions_df, base_df)
}
\arguments{
\item{predictions_df:}{A data.frame of predictions from the model.  It should have a single column of data.}

\item{base_df:}{the base data.frame for creating the predictions; 
it should have columns x and y specifying the spatial coordinates for that row.}
}
\value{


a data.frame with three columns: 'x', 'y', and 'z'.  'x' and 'y' are the spatial location of the 
prediction in the original CRS, and the 'z' column is predictions.
}
\description{
Renames columns and adds the spatial information that is lost at model inference time.
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
