% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{preprocess_raster_to_df}
\alias{preprocess_raster_to_df}
\title{a quick function based on the original code}
\usage{
preprocess_raster_to_df(raster_obj, model)
}
\arguments{
\item{df:}{A data.frame}
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
