% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{process_tile}
\alias{process_tile}
\title{processes a small raster imarge}
\usage{
process_tile(
  tile_filename,
  ml_model,
  cluster = NULL,
  return_raster = TRUE,
  return_filename = FALSE,
  save_path = NULL,
  suppress_output = FALSE
)
}
\arguments{
\item{tile_filename:}{A string specifying the location of the target raster on the disk}

\item{ml_model:}{the machine learning model for prediction.}

\item{cluster:}{a cluster, from the raster::beginCluster(); raster::getCluster() or parallel::makeCluster().  Default is NULL (no parallelism).}

\item{return_raster:}{(default: TRUE) returns a rasterLayer object if true, or a data.frame if FALSE.}

\item{save_path:}{the path to save the output.  If NULL (default) no file is saved.  Otherwise it attempts to save the file to the location specified.}

\item{suppress_output:}{if TRUE, returns the save location of the output, rather than the output itself.
If FALSE (default), the function returns a raster::rasterLayer or base::data.frame as determined by return_raster parameter.}
}
\value{

}
\description{
processes the given image in-memory.  This assumes that the image is small enough that tiling is not required.  
Functions include data munging, imputation, automated vegetation index calculation, model inference, and data type conversion. 
The outputs are optionally saved to disk as well.  
Parallization is used if a connection to a parallel (or raster) package cluster is provided
}
\examples{
Not Yet Implmented

}
\seealso{
None
}
