% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{remove_band_column}
\alias{remove_band_column}
\title{Functions returns columns that are NOT bandpasses}
\usage{
remove_band_column(x)
}
\arguments{
\item{x:}{a dataframe}
}
\value{
explanation
}
\description{
Long Description here
}
\examples{
Not Yet Implmented

}
\seealso{
remove_meta_column
}
