% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{remove_meta_column}
\alias{remove_meta_column}
\title{Functions returns columns that are bandpasses}
\usage{
remove_meta_column(x)
}
\arguments{
\item{x:}{a dataframe}
}
\value{
a dataframe without non-bandpass columns
}
\description{
Functions returns columns that are bandpasses
}
\examples{
Not Yet Implmented
}
\seealso{
remove_band_column
}
