% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{remove_noisy_cols}
\alias{remove_noisy_cols}
\title{removes columns outside the min and max indicies}
\usage{
remove_noisy_cols(df, min_index = 1, max_index = 274)
}
\arguments{
\item{df:}{A data.frame}
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
