% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{resample_spectra}
\alias{resample_spectra}
\title{Resample the data frame/spectra every distance=5 units}
\usage{
resample_spectra(spec_lib)
}
\arguments{
\item{df:}{a dataframe of spectral data}

\item{distance}{(default: 5nm): distance to resample bands}
}
\value{


data frame of resampled data
}
\description{
Long Description here
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
