% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{resample_spectral_dataframe}
\alias{resample_spectral_dataframe}
\title{Resamples the given dataframe to every 'wavelength' nanometers}
\usage{
resample_spectral_dataframe(df, wavelength = 5)
}
\arguments{
\item{df:}{a dataframe}

\item{wavelength:}{The wavelength separation for columns, in nanometers.}
}
\value{
A dataframe with the spectral compaonents
with the specified wavelength separation between bands
}
\description{
Resamples the given dataframe to every 'wavelength' nanometers
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
