% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{speclib_to_df}
\alias{speclib_to_df}
\title{Converts a spaectral library to a dataframe.  This function is not yet implemented}
\usage{
speclib_to_df(speclib)
}
\arguments{
\item{speclib:}{A spectral library to convert}
}
\value{
A dataframe
}
\description{
Converts a spaectral library to a dataframe.  This function is not yet implemented
}
