% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecospectR.R
\name{visualize_predictions}
\alias{visualize_predictions}
\title{visualizes the output of the pipeline based on a specified colormap.}
\usage{
visualize_predictions(filepath, colormap)
}
\arguments{
\item{df:}{A data.frame}
}
\value{

}
\description{
Long Description here
}
\examples{
Not Yet Implmented
}
\seealso{
None
}
